Create `processRawDiffBody` endpoint
It's better to have a separate endpoint for the playground vs. the github webhook integration because on the latter I do more involved checks, validate billing etc... Having all of this logic in a single place is just messy and too much of a hassle to deal with. It leads to bug difficult to troubleshoot